[css-flex] Try most conservative intrinsic sizing algorithm changes
In this version, we only change behavior for the most egregious cases.
We lose a lot of the benefits with this, but we need to make sure even
this simplest version doesn't have compat problems before mucking
further with the aspirational versions.

Bug: 240765
Change-Id: Ia7b4e21b10d453b847fc76ed63570025c2f3ee36
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4616797
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1160890}
NOKEYCHECK=True
GitOrigin-RevId: 46f904d16ebc598e95a3ba5b50e1b10eaa825577