Add render frame API for legacy tech report
Define the API in mojom. It will be used by the blink to pass legacy
technology event to the browser process, if the feature is enabled by
the Chrome enterprise policies.

Later on, the report will be uploaded to the enterprise service and
showing aggregated result to admins.

Bug: 1441215
Change-Id: I4976bfd9c41f3ba8ead35efef715972121b44a89
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4622108
Commit-Queue: Owen Min <zmin@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Joe Mason <joenotcharles@google.com>
Reviewed-by: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1161207}
NOKEYCHECK=True
GitOrigin-RevId: e965df231c73c08cc522d4f59bb168e43ee1cdf2