Refactor popover initial focus
This patch removes the autofocus_only parameter from GetFocusDelegate
and replaces it with a call to GetAutofocusDelegate. This removes
support for a use case where the popover is a shadow host with
delegatesFocus, but after discussing this doesn't seem like an important
use case.

This is being changed in the spec here:
https://github.com/whatwg/html/pull/8998

Bug: 1307772
Change-Id: Id7b92d467492030f4a9226ad3f5ae0a10f1896e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4336167
Auto-Submit: Joey Arhar <jarhar@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1161492}
NOKEYCHECK=True
GitOrigin-RevId: 6bfe7cd47b0de410cc4c90ffbc873ccfc870077d