[@scope] Avoid invalidation flags based on in_rightmost_compound
When matching selectors against certain pseudo-classes,
we set flags to indicate that elements depend on those
pseudo-classes. To distinguish between elements that depend on
a pseudo-class directly and elements that are *related* to such
a dependency, we use the in_rightmost_compound flag.

This does not work well with @scope and the StyleScopeFrame cache,
because we evaluate the selectors in the @scope prelude *once*,
when the scope activations are first calculated. At that time,
:scope will either be in the subject position
(in_rightmost_compound=true), *or* in the non-subject position
(in_rightmost_compound=false), and the invalidation flags will be
marked accordingly. If we're matching another scoped selector later
with :scope in the opposite position, we'll hit the cache,
and neglect to set the flags we need.

In preparation of solving this problem, this CL introduces
"impact" to SelectorCheckingContext, which controls directly how
flags should be set. The critical difference between this and
in_rightmost_compound is that it's possible to specify impact=kBoth,
which will cause *both* subject/non-subject flags to set. This will
be used by @scope in a future CL.

Bug: 1451091
Change-Id: I55b5977c452d79c7572447489d8543aa132c18df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4629312
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Byungwoo Lee <blee@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1161680}
NOKEYCHECK=True
GitOrigin-RevId: baa44f42ef9d82cea9876536ce62a6492e1f3297