Check for context loss during drawing buffer resolution.
On dual-GPU MacBook Pros on top of OpenGL, when the system activates
the discrete GPU, the WebGL context must be lost and restored in
certain situations to avoid corruption of multisampled renderbuffers.

This context loss occurs in unexpected places, in particular during
execution of the ScopedDrawingBufferBinder. Check for it and early-out
in more places. Add a regression test which catches the bug.

Bug: 1441633
Change-Id: Ib1ce1c7e11e5180edd63430180c461789bf766fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4639167
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Commit-Queue: Kenneth Russell <kbr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1161934}
NOKEYCHECK=True
GitOrigin-RevId: df92771d95964b74ea477cfd89512f48d6d1abf0