DevTools: allow running addScriptToEvaluateOnNewDocument for existing contexts
Currently, during auto-attach to a target while the target is paused there is no way to run a script for an isolated world that will be created once the target is unpaused. Runtime.evaluate requires a contextId which can only be retried after the target is unpaused. This CL adds the flag to opt-in into running the script for "existing" documents immediately for the following reasons:

1) worlds is a Blink feature and should not be exposed in the Runtime domain (V8 feature).
2) there is a precedent in addScriptToEvaluateOnNewDocument for running the scripts for paused new windows already.

This parameters allows running the script immediately even if the target is not paused and while it is not required for the current issue with auto-attach and isolated worlds it might be useful on its own because addScriptToEvaluateOnNewDocument calls are often followed by the Runtime.evaluate calls with the same script.

Fixed: 1449849
Change-Id: I0ddb535bd5c283f6859018eaea7621712954649f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4637325
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1162267}
NOKEYCHECK=True
GitOrigin-RevId: 267dbb8b5f6604eb9b76efea2764794e90ffdd5c