Inline CSSToLengthConversionData::FontSizes() constructors.
These are trivial, so can just as well be in the .h file. The perf gain
is small enough that it's hard to measure the usual way, but seems to be
about 0.4% if looking at the UpdateFont() relative times in the profiler.

Change-Id: I6198c2e08a0b760b3293289c76cdf31a6e48460e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4641111
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1162324}
NOKEYCHECK=True
GitOrigin-RevId: d4b2b1936784bf87a3ed54530651b764aa8fd334