Only match @starting-style rules while resolving style
StyleResolver/ElementRuleCollector has different collection modes too,
for which we shouldn't include the @starting-style rules.

In particular, devtools showed @starting-style rules as matching,
overriding declarations in rules for after-change style. Since
@starting-style rules are immediately overriden as part of the style
change event, which means it doesn't make sense to expose these as
matching rules to devtools for the time being.

For transitions not involving starting-style, the matched rules in the
devtools panel only shows the after-change style rules while a
transition is running. If that is ever changed to show both from and to
style rules during a transition, it would also make sense to expose
@starting-style rules affecting the from style.

Bug: 1412851
Change-Id: I548ce8ea7da623710da80a68c3296875a3aa1486
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4640128
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1162362}
NOKEYCHECK=True
GitOrigin-RevId: 3ecba148bd1792b1f15b599fbb4a0f767d63b33a