Make use of Skia's codecs explicit in Blink gif decoding and viz unittests
In http://review.skia.org/689016, Skia added APIs to allow a single
codec to be explicitly used.

This CL replaces the 2 use of SkCodec::Make* with an explicit,
specific codec that is intended. This means it will no longer
try to decode the bytes with all of the compiled-in Skia codecs,
but only PNG or GIF (depending on the file).

Bug: skia:13983
Change-Id: Ia58061be43027c4b2f6b4eecdf6dc6ea8357c61d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4615536
Reviewed-by: Florin Malita <fmalita@chromium.org>
Commit-Queue: Kevin Lubick <kjlubick@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Reviewed-by: Leon Scroggins <scroggo@google.com>
Cr-Commit-Position: refs/heads/main@{#1162373}
NOKEYCHECK=True
GitOrigin-RevId: 61afceadc4c747b0de775aeb25aeac5500708ae8