Implement get_fedcm_dialog_title and select_fedcm_account in TestDriver
This CL implements two of the FedCM automation commands from
https://fedidcg.github.io/FedCM/#automation. This CL does not yet update
the tests to make use of these methods. A followup is required to stop
auto-selecting accounts in tests so that we can reliably get the title
before the UI is dismissed.

This is needed because WPTs run under content_shell, without
chromedriver support.

Bug: 1453691
Change-Id: I0f7765aa012b38d6e5ce73070cb47341af4f57a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4618303
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Nicolás Peña <npm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1162433}
NOKEYCHECK=True
GitOrigin-RevId: 9be3a4cb435b0c20d1d6634e1ec986a204153c81