Add DevTools command to set Attribution Reporting local testing mode
By asynchronously recreating the AttributionStorageDelegate with the
new value.

This will be used as the implementation for forthcoming web driver and
test driver commands to allow Attribution Reporting WPTs to run quickly
and deterministically by disabling noise and report delays.

In the future, we will augment the Storage domain with additional
commands and events for interacting with the Attribution Reporting API
in order to deprecate and remove the corresponding internals UI.

https://github.com/WICG/attribution-reporting-api/pull/843

Bug: 1457165
Change-Id: I69f4620b4c389fe72d621942654073c491f230a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4632317
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1162911}
NOKEYCHECK=True
GitOrigin-RevId: 953b155a02e0a21902fcd6a98788d938d70ce0cc