Implement ComputeInteger for css math nodes and sign()
In order to be able to use deferred calculation for sign() function
we need to have a ComputeInteger with css resolve data to give this info
to the css math node.

For testing purposes only z-index will work with integer resolution for
now.

Design doc:
https://docs.google.com/document/d/1rikPXstHFK3b8N95QzfDar6dDDrEsuVgw1YF-JgvKQM

Bug: 1436612

Change-Id: I574e27d84e37b7a11397b0c0f14f3a183648d795
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4607246
Commit-Queue: Daniil Sakhapov <sakhapov@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1162938}
NOKEYCHECK=True
GitOrigin-RevId: eabeca9590f1b2314d98367fc3bce445a6459ccf