Rename blink::UrlPattern to blink::SafeUrlPattern.
This is a follow up CL of:
https://chromium-review.googlesource.com/c/chromium/src/+/4570363/comment/6ce953c7_50911e39/

There is blink::URLPattern, which is used for the URLPattern API.
https://developer.mozilla.org/en-US/docs/Web/API/URLPattern/URLPattern
It should be used within blink to realize the API.

At the same time, there is blink::UrlPattern.  It was renamed from
blink::Manifest::UrlPattern in:
https://chromium-review.googlesource.com/c/chromium/src/+/4422961
The renaming happened to make URLPattern usable not only Manifest but also
ServiceWorker Static Routing API.
https://github.com/yoshisatoyanagisawa/service-worker-static-routing-api

In both cases, blink::UrlPattern aims to bring URLPattern definition in
the renderer process to the browser process to make the evaluation
executed there.  Therefore, its capability is a bit limited due to the
security reason.  Especially, a user-defined regex is not allowed.

Considering the difference, I believe it is natural to leave blink::URLPattern,
which might need to be the same name with the API, and changing
blink::UrlPattern to blink::SafeUrlPattern to clarify it is
modified for the security reason.  To allow URLPattern usable in
the browser process, user-defined regular expression has been
dropped for SafeUrlPattern.

Bug: 1371756
Change-Id: I719366a4e021ccf8441fab5ee49ecec2f3378ff4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4617993
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Commit-Queue: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1163349}
NOKEYCHECK=True
GitOrigin-RevId: 91be4bcad0a436fd0fd1ef5c8a1000c8969af137