WebNN: Enable pthreadpool for MLGraphXnnpack
Thanks to the pthreadpool's integration with Chromium Jobs API [1], this
CL creates a pthreadpool for XNNPACK Runtime that enables parallel
execution of XNNPACK operations with `base::ThreadPool` workers. Because
the `struct pthreadpool` is accessed by XNNPACK Runtime without a lock,
a `SequenceChecker` is used to ensure the accessing is thread-safe.

This CL experimentally allows the JavaScript code to set the number of
parallel workers via `MLContextOptions.numThreads` when creating an
`MLContext` of XNNPACK backend. The default value 0 means no parallel
execution. The user-supplied threads number would be capped by the
minimum value of the number of logical processors and `kMaxNumWorkItems`
which is set to 4 according to the initial testing result.

This CL also adds a dedicated `MLGraphXnnpackTest.ThreadPoolTest` that
tests the creation and execution of `MLGraphXnnpack` with different
number of threads.

[1]: https://source.chromium.org/chromium/chromium/src/+/main:third_party/pthreadpool/chromium/jobs.cc

Bug: 1273291
Change-Id: Ide652eb3cc90760487628ff680009681d45f4ec1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4485367
Reviewed-by: Jiewei Qian <qjw@chromium.org>
Commit-Queue: ningxin hu <ningxin.hu@intel.com>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1163411}
NOKEYCHECK=True
GitOrigin-RevId: 5115b843db204cf420f799c698fd5d40dde9e65d