Fix OoB read in neon implementation of ProcessARateVectorKernel
The test case and reproduction steps are attached to the bug report.
The short version is: a relatively simple JS only test case makes it
possible to hit the following DCHECK on Linux/armv7. In non DCHECK
builds this results in OoB access to two arrays.

[3664:3676:0627/114720.581346:8348136604:FATAL:oscillator_kernel_neon.cc(219)] Check failed: read0[m] < periodic_wave_size (4096 vs. 4096)
0  0x000003f90886 base::debug::CollectStackTrace()
1  0x000003f81c14 base::debug::StackTrace::StackTrace()
2  0x000003ed81b2 logging::LogMessage::~LogMessage()
3  0x000003ec4db6 logging::(anonymous namespace)::DCheckLogMessage::~DCheckLogMessage()
4  0x000003ec4c98 logging::CheckError::~CheckError()
5  0x00000611b36e blink::OscillatorHandler::ProcessARateVectorKernel()
6  0x00000610c7de blink::OscillatorHandler::ProcessARateVector()
7  0x00000610c95e blink::OscillatorHandler::ProcessARate()
8  0x00000610cbec blink::OscillatorHandler::Process()
9  0x0000060dcc30 blink::AudioHandler::ProcessIfNecessary()
10 0x0000060e1e4e blink::AudioNodeOutput::Pull()
11 0x0000060e18a8 blink::AudioNodeInput::SumAllConnections()
12 0x0000060e19aa blink::AudioNodeInput::Pull()
13 0x000006111a68 blink::RealtimeAudioDestinationHandler::Render()
14 0x00000611290e blink::AudioDestination::RequestRender()
15 0x0000061125fc blink::AudioDestination::Render()
16 0x00000658280c content::RendererWebAudioDeviceImpl::Render()
17 0x000006582862 content::RendererWebAudioDeviceImpl::Render()
18 0x000001d23960 media::SilentSinkSuspender::Render()
19 0x000001cfe7d0 media::AudioOutputDeviceThreadCallback::Process()
20 0x000001cf56c8 media::AudioDeviceThread::ThreadMain()
21 0x000003f74934 base::(anonymous namespace)::ThreadFunc()

I'm far from an expert when it comes to arm/neon, or webaudio, but
from what I can tell the code already expects the input values can be
outside of periodic_wave_size bound. This is why the read_index_mask
is applied to read1 array. It does not get applied to read0 for which
the DCHECK fails. I think this is the main bug here. In fact when
looking at SSE2 implementation of ProcessARateVectorKernel the
read_index_mask is applied to both read0 and read1. Doing the same for
arm neon implementation fixes the DCHECK and OoB read in release builds.

Bug: 1458355
Change-Id: I1b78321e57e97f0f2fd117ece0bb5a8ea4bed8e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4648650
Reviewed-by: Michael Wilson <mjwilson@chromium.org>
Commit-Queue: Piotr Tworek <piotr.tworek@xperi.com>
Cr-Commit-Position: refs/heads/main@{#1163421}
NOKEYCHECK=True
GitOrigin-RevId: fae10d98beda1233d6e9c643d6cf8b0b76cefb36