Send legacy tech report from blink
Deprecation::CountDeprecation will send the report to a new end point
that is located at frame.mojom API. It will then later upload to
enterprise management service through the browser process.

The event will only be sent when the feature is enabled by an enterprise
policy so that we avoid unnecessary IPC for normal users. A command line
flag is used to pass the flag to the renderer process.

Bug: 1441215
Change-Id: Ic9a2b3c4a4e1c8de121dba1e102372a3940a82fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4641912
Reviewed-by: Ari Chivukula <arichiv@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Owen Min <zmin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1163513}
NOKEYCHECK=True
GitOrigin-RevId: cd281b7a50555bc0e430a39840931de7b7c55bf8