Allow select to have writing mode vertical
This CL allows <select> to have writing-mode vertical for both the
default native appearance and when appearance is set to none.
The select arrow direction changes based on vertical-lr or vertical-rl.
The select arrow position changes based on direction LTR or RTL.

We also make changes to support the <select multiple> case
- overflow-* is based on block/inline instead of x/y.
- arrow keys left/right select the correct option.

We also put the change behind the feature flag
FormControlsVerticalWritingModeSupport so it can be turned on safely.

Change-Id: I5d3ff770c1933667238fe9352fe562f85a541f3a
Bug: 681917
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4505459
Commit-Queue: Di Zhang <dizhangg@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1163589}
NOKEYCHECK=True
GitOrigin-RevId: a8cdabae8b2decf0c6b222b1f6e5803208abd806