Create a directory for //third_party/mediapipe
MediaPipe will need many patches in order to be made compatible with
Chrome, so we will need to land the files directly in the src tree.

This CL establishes a new directory with OWNERS and a script that will
be used to bring in the 3p source.

There will be several CLs in fast follows to add 500 files at a time,
to reduce the burden on Gerrit since we need to add ~1,500 files in
total.

This addition has been approved by ATLs and Security.

Bug: b/283522287
Change-Id: I314c0fd218b26f6c5c14480742a91d00f66cfa65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4652003
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Commit-Queue: Robert Ogden <robertogden@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1163609}
NOKEYCHECK=True
GitOrigin-RevId: 3a8da1da64a0a6557827f7286ef5e8e01bb9fa8f