Honor @page size and margins in web tests.
After layout for pagination, check if the size of the first page is
different from what was specified initially. If it is, relayout with
that page size. This is the same as what PrintRenderFrameHelper does
(i.e. what we do in regular Chrome). However, we will need to rewrite
this when we add support for mixed page sizes, but before we get there,
it's a step in the right direction not to hardcode the page size, and
behave more similarly to regular printing in Chrome.

This should have allowed us to set the default WPT page size to 5 by 3
inches, just like
http://web-platform-tests.org/writing-tests/print-reftests.html says,
but that would break infrastructure/reftest/reftest_match-print.html
(which passes in Gecko), so we need to keep on using 4 by 2 inches for
now.

Update the bug number for one of the test failures, since it turned out
to be unrelated to this issue.

No behavior changes outside running web tests intended.

Bug: 1416181
Change-Id: I809901af7c2a1f8b06947de997dd226722fe8ee7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4654234
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1163637}
NOKEYCHECK=True
GitOrigin-RevId: a87b8d076f097e476b3f900a322d07e396fb740f