Upload 500 unpatched files of //third_party/mediapipe
MediaPipe will need many patches in order to be made compatible with
Chrome, so we will need to land the files directly in the src tree.

There will be several CLs in fast follows to add 500 files at a time,
to reduce the burden on Gerrit since we need to add ~1,500 files in
total.

Bug: b/283522287
Change-Id: Iea504ed14d737e4f99c4d52cb0133b0b4f85f138
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4654795
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Commit-Queue: Robert Ogden <robertogden@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1163801}
NOKEYCHECK=True
GitOrigin-RevId: 4c43a7157b3c1cf8b430d1c6e6288b16fcf3d834