[GridNG] Fix DCHECK with invalid subgrid auto-repeaters
This adjusts the conditions for calculating subgrid auto repeats so we
don't go down the non-subgrid path when `subgrid` is specified without
a valid subgrid span (which is necessary to compute repetitions).

Ethan brought up an interesting question which prompted me to adjust
our computed style behavior to match the spec. A new test was added
to validate this behavior.

Bug: 1457592
Change-Id: If63b1f3963b7833d2bdad9b3bd573016df3c61a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4647952
Reviewed-by: Ethan Jimenez <ethavar@microsoft.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Kurt Catti-Schmidt <kschmi@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1163818}
NOKEYCHECK=True
GitOrigin-RevId: 193f1a670c6ba3e008a2cf6876366b555310cf8f