Upload up to 1,000 unpatched files of //third_party/mediapipe
MediaPipe will need many patches in order to be made compatible with
Chrome, so we will need to land the files directly in the src tree.

There will be several CLs in fast follows to add 500 files at a time,
to reduce the burden on Gerrit since we need to add ~1,500 files in
total.

Bug: b/283522287
Change-Id: Id46c773d89b62202ccbe324a4e443a62d1911e33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4652064
Commit-Queue: Robert Ogden <robertogden@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1163835}
NOKEYCHECK=True
GitOrigin-RevId: 322d3952e0df0cad578a62ca7aecbf702a7a5e23