Remove the Finish*() member functions from MatchResult.
This simplifies the interface and makes for less implicit state.
The grand plan is to try to simplify things enough that we can
skip MatchResult altogether and move to running ExpandCascade
immediately when we get back rules.

It is possible that we should also get rid of
BeginAddingAuthorRulesForTreeScope() for the same reasons,
but this patch leaves it unchanged for now.

Change-Id: Icde422202db3000ea6ec47a2b9fa96806dd2d990
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4652009
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1164066}
NOKEYCHECK=True
GitOrigin-RevId: f4c890d7e6f986582cdbac2ca209866022d4b588