Revert pixel snapping scale-adjustment behavior for SVG roots
This reverts the main behavioral change from the change in commit
crrev.com/d1d05d02de7519c777358ce926c007a7222449d6 . This bug has seen a
lot of duplicates, which suggests that it is non-obvious which of the
two bugs is worse. A flag is added to gate this in case it turns out
that this new (old) cause significant breakage. Symptoms of the bug
fixed by crbug.com/599095 is "dropouts" along the right/bottom edges.

Only revert to the pre-599095 behavior for inline SVG fragments, since
in other cases we'll be rendering into integer sized viewports.

The test from the fix for issue 599095 [1] is marked as failing.

[1] svg/custom/zoomed-background-alignment.html

Bug: 848748, 599095
Change-Id: I8ef1d61e30d6bb72b2dc1b2fba091215e8c42262
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4644175
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1164068}
NOKEYCHECK=True
GitOrigin-RevId: 152bac596c24b9c71963b50ae19c36435b676083