[@scope] Propagate CSSNestingType instead of in_nested_style_rule
In CSSSelectorParser, we currently propagate a boolean
in_nested_style_rule as a parameter, and separately store
the CSSNestingType on CSSSelectorParser itself. We can merge these two
into one, as the nesting type is only relevant when we're in a nested
style rule.

This is just a simplification to make future CLs more palatable,
and there should not be any behavior change.

Bug: 1456435
Change-Id: Ibbceb879715fab7055c9aa644e11541d3d816ff3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4650390
Reviewed-by: Steinar H Gunderson <sesse@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1164079}
NOKEYCHECK=True
GitOrigin-RevId: 7a4c97befc90d245e5db05445b742072ca1ea261