[scheduler] Fix UAF in sequence_manager_fuzzer
Prior to crrev.com/c/4356316, the post-delayed-task action in the
fuzzer held a scoped refptr to the target TaskQueue, ensuring the queue
stayed alive while posting the task. Changing this to a raw pointer
could cause a UAF if racing with shutdown on the owning thread. This
CL fixes this by synchronizing accessing the target task runner with
shutdown. Note that alternatively it could hold a scoped reptr for
TaskQueueWithVoters, but eventually shutdown will be replaced with
just clearing the queue pointer, which would could have the same
problem.

Bug: 1143007,1458376
Change-Id: Ic7a9db1dd860be3d02cd1bed5cd2ed55361b336e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4652937
Reviewed-by: Francois Pierre Doray <fdoray@chromium.org>
Commit-Queue: Scott Haseley <shaseley@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1164245}
NOKEYCHECK=True
GitOrigin-RevId: 4b7faeaa9c149aeb1c700d8aeea1c21917f5893f