[anchor-position] Fix 'anchor-scroll' with fixed-positioned scrollers
When a layer is fixed-positioned, its parent layer is LayoutView but
it doesn't scroll with it, which requires some fixups.

AnchorScrollData already performs the fixup when calculating the
scroll container layer for the anchor and the target elements, but
misses to do so when walking up the layer tree. This patch fixes it.

Fixed: 1459079
Change-Id: Ia0fd01fa2f21e020fc1ff70b796e8e188489b27e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4655618
Auto-Submit: Xiaocheng Hu <xiaochengh@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1164259}
NOKEYCHECK=True
GitOrigin-RevId: 3bd92bd77eb220d2e28e3990963a4ab0c6332a1c