Invalidate AX objects from table cells if their role might change
The role of a table cell depends on its siblings in a complex way
(see DecideRoleFromSiblings), so all siblings need to be invalidated
if any of them change their text content or are added.

In addition to the newly added test, this CL deflakes
DumpAccessibilityTreeTest.AccessibilityTableHeadersOnAllSides,
which sometimes fails on the bots for CL 4553316 due to timing
race conditions caused by the HTML parser.

Bug: 1342801

Change-Id: Ia9119db8d2db7efde5775acca2f46ca9a737e48e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4641319
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Commit-Queue: Chris Harrelson <chrishtr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1164333}
NOKEYCHECK=True
GitOrigin-RevId: 1559a168c04ba03f020e9c88294ef5088c035876