[CT] Send initial scroll % signal
With this CL, RealtimeEngagementSignalObserver will send an
onGreatestScrollPercentageIncreased signal immediately after it is
created if the scroll percentage is > 0. This won't make a difference
until RealtimeEngagementSignalObserver's creation is deferred until an
EngagementSignalsCallback is set.

On the blink side, this CL updates RenderFrameMetadataObserver to push the RenderFrameMetadata to the browser if the RootScrollOffsetUpdateFrequency increased or was set for the first time.

Bug: 1340074
Change-Id: I691a5058f0bf79b5f443312f7305a14032ca28d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4544201
Commit-Queue: Sinan Sahin <sinansahin@google.com>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Reviewed-by: Bo Liu <boliu@chromium.org>
Reviewed-by: Florin Malita <fmalita@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1164429}
NOKEYCHECK=True
GitOrigin-RevId: 74c972bdf2bbf74eec8aa6a30d3c1e993eb93ef4