InlineLayout: add initial support for inline text-box-trim
This CL implements the initial support for the text-box-trim property.
More specifically, it makes blink calculate the half-leading of inline
boxes based the value of its text-box-trim property.

If the property asks to trim the half-leading from the start/end side
of the box, half-leading will be treated as 0, to match the text-over baseline/text-under baseline.

Design doc:
https://docs.google.com/document/d/1eT0xEOrfVI3RhZ6goJLW1yOWHXLBq2tzAYrHe3NIiO8/edit

Bug: 1411581
Change-Id: I3c7f5a1bfcdfe75b2b84a73ebdc81e53e973d69a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4545925
Commit-Queue: Lingqi Chi <lingqi@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1164490}
NOKEYCHECK=True
GitOrigin-RevId: f90c336c1c99af4e80010d3991449ff66e6efc26