[RemoveLegacy] Remove NeedsPositionedMovementLayout flag.
This flag was used when an inset/margin changed on non-static
positioned elements.

In NG we didn't skip the LayoutResult cache if it was marked as dirty
with this flag, and legacy performed a special type of "movement"
layout when dirty.

We don't need this flag anymore, instead we can directly invalidate
the containing block (needing "simplified" layout if the element is
OOF positioned, or "normal" layout if not).

This patch will likely result in some elements not requiring full
layout anymore. E.g. DependenceOnContentHeightHasChanged triggered a
full layout on a rel-pos element when not required.

For rel-pos we could always just require "simplified" layout if we
undid the rel-pos offset, then recalculated it correctly. There is
some complexity with %-insets under this scheme however.

Bug: 1229581
Change-Id: I409195ee5aa2e14cd2004b0bfca43485cd5f64db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4655244
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1164753}
NOKEYCHECK=True
GitOrigin-RevId: 4d1fb0d1b5b2e173b9d432b884d51ba170f7e383