wei: Run only for android and add mojo mocks
This feature relies on platform features and is going to be
probabilistic. The way contacts-manager wrote tests is to have
a standard expected mock contract that each browser needs to
fulfill and then mocks out for the mojo calls so following that
standard while we work on this.

The getEnvironmentIntegrity mojo call currently isn't used for
anything so just including a test that verifies the mojo call
resolves until we can use this in tests.

We can consider re-enabling this for other platforms in the
future because of our mocks but the mojo remote is only
set up for android currently.
As such, disabling this for all other platforms besides android.

As such, we can also update these tests to check that the API
can be called. It was previously checking for failure because
it was being assumed it was being run in a Linux environment.

Test: run_wpt_tests.py
Bug: 1454404
Change-Id: If6a67f808f00d18b21bb809594bb2537e73c77af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4609130
Auto-Submit: Rupert Wiser <bewise@chromium.org>
Commit-Queue: Weizhong Xia <weizhong@google.com>
Reviewed-by: Weizhong Xia <weizhong@google.com>
Cr-Commit-Position: refs/heads/main@{#1164772}
NOKEYCHECK=True
GitOrigin-RevId: d7c79ff725db2ead4fabae71d938d2bb565e2054