[a11y] Serialize text deletion/insertion offsets for accessibility
This CL adds logic to notify accessibility of changes in text range's
offset with respect to their anchor nodes.

This is to fix a UIA bug where the text offset in
the `TextRangeProvider` is out of sync when a deletion is made, which
then leads to some ATs that expect the text range to be synched with
the renderer to have wrong behavior.

The way this is done in this CL is:
1) When a deletion is made through a backspace command, we notify the
`AXObjectCache` of the deletion, if the cache already exists.
This is also done if an insertion happens.
2) The `AXObjectCache` then sends a tree update with information
on where the deletion/insertion has happened. This information
is sent through new IntList attributes that were added in this CL.
3) On the browser side, we notify observers when this is
received, and these update the offset accordingly when necessary.

Change-Id: I6c12d9816292d00771c8a4afacde65e04169061c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4563035
Commit-Queue: Javier Contreras <javiercon@microsoft.com>
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1164871}
NOKEYCHECK=True
GitOrigin-RevId: 36441f97ee3931d4aa62f27e7e2aa05cc8b20764