Make LayoutSVGContainer drive transform updates
Drop the `bounds_changed` argument passed to `CalculateLocalTransform()`
and handle the potential need for implicit transform invalidation in
LayoutSVGContainer::UpdateLayout().

Make only LayoutSVGContainer handle the `needs_transform_update_` flag,
and make the transform update itself the only responsibility of the
various subclasses.

Rename LayoutSVGContainer::CalculateLocalTransform to
UpdateLocalTransform to better reflect what it does - since it has
side-effects.

Move the transform invalidation for 'x' and 'y' on <use> to style update
time, and only update it in the corresponding transform update hook. Add
a case to SVGContentContainer::Layout() to handle when the viewport
itself changes and 'x' and/or 'y' is a percentage.

Bug: 924472
Change-Id: I4922cb5d76dded9f9d36540a24a6d806f183e1b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4660615
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/main@{#1165813}
NOKEYCHECK=True
GitOrigin-RevId: 42f6a920dcab3d9753e0cd681d77c94ba717d112