Get rid of EnsureComputedStyle from DOM snapshot code
Avoid EnsureComputedStyle for Node as we are about to only expose it on
Element. The existing check for LayoutObject means the element is
rendered and the style should exist on Element and the Ensure* call
should be unnecessary.

The check for LayoutObject is dubious due to the display:contents case,
but this CL tries to keep the existing implementation behavior. Left a
comment to say so.

Bug: 1459564
Change-Id: I2893fe98e38f966482e9c8b72c846ae089fba875
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4665404
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1165914}
NOKEYCHECK=True
GitOrigin-RevId: 3a408ea5e969726e38dbd26ce838a09449f31916