Explicitly pass the reference box to SVG transform calculation
Since computing the stroke bounding box - as used for
'transform-box: stroke-box' - requires the children of a container (or
similar) to first have their layout updated, it's no longer possible to
call SVGElement::CalculateTransform() since that computes the reference
box itself. Instead hoist that computation out into the callers that
need it.

For LayoutSVGBlock add a new UpdateTransformBeforeLayout() (symmetric
with the existing UpdateTransformAfterLayout()) and use that in the two
subclasses. In other cases just rely on the shared implementation in
TransformHelper, and explicitly compute and pass in the reference box.

Bug: 924472
Change-Id: I3e8d2549f2b89534d1f544618a0c3f0c290a1bef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4661019
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1165974}
NOKEYCHECK=True
GitOrigin-RevId: a78a62e87ff8c9cf9cccbaf9b9f881999098ed87