Enable cardboard runtime to build by default
When this initially landed building was disabled by default as the
implementation wasn't even far enough along for a reasonable runtime
flag to guard the feature. While still missing some parity with the
existing GVR runtime, the implementation is at least at a reasonable
point to include it in the build to ease testing and development.

The cardboard runtime increases the size of the binary by 76kb; but
this is actually smaller than the ARCore runtime and long term will
enable us to remove the VR DFM, and will improve the reach and
performance of WebXR/VR overall.

Given that the default conditions for enabling all of the WebXR Android
supported runtimes are the same, this change also abstracts that to
a separate, more easily modifiable arg that the other parameters all
reference. The intention here is to provide some re-use, while also
allowing other embedders the ability to toggle specific runtimes on or
off via overrides.

Bug: 989117
Binary-Size: See Above
Change-Id: Iffef1073b202220d9cd1783d7f298254eb21e99e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4655154
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Commit-Queue: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1166026}
NOKEYCHECK=True
GitOrigin-RevId: b4e234ee9a89b2cf861b044ebe583118bfd41d70