[blink-tests] Update iterator blink tests
This CLs updates blink web and wpt tests to be compatible with
shipping iterator helpers.

The iterator helpers proposal adds new methods to the Iterator
prototype and exposes it as `Iterator` on the global. It also
changes the value of [Symbol.toStringTag] from [object Object] to
[object Iterator]. These changes are tested in V8 and test262.

Bug: v8:13558
Change-Id: Ia6805f9f93ff3fb534958471c1b57ada6ba0b662
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4655610
Reviewed-by: Dominic Farolino <dom@chromium.org>
Commit-Queue: Shu-yu Guo <syg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1166091}
NOKEYCHECK=True
GitOrigin-RevId: d6179b77f20f09285696a2d2c5bf999193668cbf