[Blink] Guard VideoFrame's per-plane readback of multiplane SI
Just as it is not possible to do per-plane readback with legacy SI if
the number of textures != the number of planes, it is not possible to
do per-plane readback with multiplane SI if doing external sampling. In
both cases we are representing the image by a single *opaque* texture
on the service side. This CL changes webcodec/video_frame.cc to
detect the second case and handle it in the same way as the first case.

Bug: 1457134
Change-Id: Iafd7513e9902100b37ca909664dd46fa723369c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4661002
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1166364}
NOKEYCHECK=True
GitOrigin-RevId: a3636d1ffd79769d79cb736dcf0066eeef408762