Move CompressionDictionaryTransport related feature flags to network
This CL moves the `kCompressionDictionaryTransportBackend` and
`kCompressionDictionaryTransport` feature flags from blink to network.

This is intended to expose the feature state to the network service.
(The `blink` feature flags are not accessible from the network service.)

During the Origin Trial experiment of the CompressionDictionaryTransport
feature, we want to have a maximum expiration time on dictionary entries
to keep the duration constrained. To know that the feature is enabled
by an Origin Trial token, the network service must be able to check the
state of the `features::kCompressionDictionaryTransportBackend` flag.
(If `features::kCompressionDictionaryTransport` is disabled but
`shared_dictionary_writer_enabled` of the request is true, this means
that the feature is enabled in the Blink side by an Origin Trial token.)

Bug: 1413922
Change-Id: I3cfef4fd112ee229a44a17819c320d7c9d4dcfd5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4664354
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1166376}
NOKEYCHECK=True
GitOrigin-RevId: b697e7b87ed048ecfed1b879f64844f6d99e6865