Stop reading from the provisional ComputedStyle before MPC lookup.
After looking up in the MatchedPropertiesCache, we call IsCacheable()
to check whether we can use the cache or not. However, this function
is really meant to check the cacheability of the finished ComputedStyle,
and  we call this on a provisional ComputedStyle, where we only really
care about the TextAutosizingMultiplier and the other fields are
at the default.

Thus, move the TextAutosizingMultiplier check into Find(), and don't
use the provisional ComputedStyle at all in this spot (we store the
autosizing multiplier in the StyleResolverState instead of in the
ComputedStyle). This is another step on the way to simply not having
ComputedStyle until we've been through the MPC lookup.

Change-Id: Ibc977b74431c84d1dd9e01b93497de599c9eecf9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4667265
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1166406}
NOKEYCHECK=True
GitOrigin-RevId: 66cc5c53174de70732f58369dd4bc68808c192a2