[css-flex] Fix arithmetic underflow causing ~infinite flexbox width
Turns out ((wtf_size_t)0 - 1) is a very big number.

We were adding gap*(num_items - 1) to the intrinsic with of flex
containers, but only when the container has a child that isn't a flex
item, like an abspos child. When the container had 0 children we
weren't running any of this logic at all as an optimization.

Fixed: 1459336
Change-Id: Ia060c6745de228598af9ec8c214ba79b5b575edc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4668132
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: David Grogan <dgrogan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1166613}
NOKEYCHECK=True
GitOrigin-RevId: a08b66c9a697427fd9f9a9063ea4f3cf700ee752