Allow input type date etc to have vertical writign mode
This CL allows <input> types "date", "datetime-local", "month", "time",
"week"  to have writing-mode vertical for both the default native
appearance and when appearance is set to none.

We also put the change behind the feature flag
FormControlsVerticalWritingModeSupport so it can be turned on safely.

Note that, for direction RTL, the calendar picker (for both horizontal
and vertical writing mode) is currently only on the right side. That
change will be tackled by issue 1432009.

Change-Id: Ib1aa557109e234ae697a15b5e8cb4eff1360a5d6
Bug: 681917
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4652558
Commit-Queue: Di Zhang <dizhangg@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1166732}
NOKEYCHECK=True
GitOrigin-RevId: 2fe598e8585c54d0cbfb6a5ee4554d79f8d6e3e0