[layout] Fix relayout boundary logic.
<div id="a">
  <div id="b">
    <div id="inner"></div>
  </div>
</div>

Where "a" and "b" are potential relayout boundary targets.

Previously it was possible to have the following invalidation logic:
1. "inner" changes - and invalidates layout up to "b".
2. "b" gets added as a relayout boundary.
3. "b" changes its *own* style - and invalidates layout up to "a".
   (note this may disqualify "b" as a relayout boundary).
4. Layout starts at "b" but potentially with the wrong constraints.
5. Layout starts at "a", but doesn't revisit "b" as it is layout clean.

This caused various types of DCHECKs and potentially incorrect layouts.

This patch ensures that we don't relayout as a root if we are marked as
needing layout by our *own* style.

Fixed: 1462742
Fixed: 1454351
Change-Id: I0ff0818ac85844acf5f6ec88ad73db83147ee01e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4671128
Reviewed-by: David Grogan <dgrogan@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1167704}
NOKEYCHECK=True
GitOrigin-RevId: d5a08f0a0db3861630f7a082075d5647a41f8436