Use HandleKeyboardActivation rather than duplicating its code.
This chunk of code originated in HTMLButtonElement.  In
https://crrev.com/8ebbf509126641fea90b694a58a3f98689d94cb8 it was copied
from HTMLButtonElement into HTMLSummaryElement (with no indication of
copying left in the code).  In
https://crrev.com/93823ec8ea86211c97bf16a1e8bd0285b3a20189 the original
HTMLButtonElement version was refactored into
HTMLElement::HandleKeyboardActivation so that it could be shared with
HTMLPortalElement.

This change makes HTMLSummaryElement share that refactored version
rather than maintaining its own copy.  Thankfully, the code is still
identical, other than the refactoring that was done to make it a
separate method.

There is also a more heavily refactored version of this in
BaseCheckableInputType and in KeyboardClickableInputTypeView that I'm
not trying to refactor here.  The former was refactored out of
HTMLInputElement in
https://crrev.com/1e961c5625c998a178a3addc6ca60596e8c4b6b3 and the
latter was copied, likely from the former, in
https://crrev.com/4ac48dd80b65c4b827dee77a3562a2f414d0fe85.  (Both the
original HTMLButtonElement and original HTMLInputElement versions mostly
originated in
https://crrev.com/689f576a65ae1a10943ba98b10621dec830d1e52.)

(In the future I would like to fix some bugs in this code, so it's
useful to have it in fewer places.)

Bug: 1361983
Change-Id: Icb1e4b6ee6135a10f9f41ecd34a9cdba220a19aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4673631
Auto-Submit: David Baron <dbaron@chromium.org>
Commit-Queue: David Baron <dbaron@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1167881}
NOKEYCHECK=True
GitOrigin-RevId: 35abdc6545c8bbbdf6be94538ebe69039d884798